refactor: Split ImageGenerationViewModel into concern-based extensions
Decompose the view-model god object (2057 -> 765 lines) into three
behavior extensions, leaving the core type to own state and wiring:
- +Selection: mask-tool enablement, layer commits, session undo/redo
- +Projects: new/open/save (bundle + legacy), startup/last-project restore
- +Generation: T2I/I2I/fill/outpaint lifecycle, saving, presets, checkpoints

Mechanical extraction; visibility widened only where extensions need
cross-file access. No behavior change.
