+++ Sources/Flux2App/ViewModels/ImageGenerationViewModel.swift
@@ -82,7 +82,7 @@ class ImageGenerationViewModel: ObservableObject {
     /// The enhanced prompt when upsampling is on, surfaced live (set the moment
     /// the VLM returns it, before the long denoise) and cleared at the start of
     /// each run. `nil` whenever the prompt wasn't upsampled.
-    @Published private(set) var upsampledPrompt: String?
+    @Published var upsampledPrompt: String?
 
     // MARK: - T2I Parameters
     @Published var width: Int = 1024
@@ -112,9 +112,13 @@ class ImageGenerationViewModel: ObservableObject {
     /// Resolved Vision silhouettes keyed by mask-layer id (preview + generate).
     @Published var visionSubjectMasks: [UUID: CGImage] = [:]
     @Published var visionSubjectStatusMessage: String?
-    private let selectionUndoStore = SelectionUndoStore()
+    let selectionUndoStore = SelectionUndoStore()
     let editHistoryStore = EditHistoryStore()
-    private var isApplyingSelectionUndo = false
+    var isApplyingSelectionUndo = false
+    #if canImport(AppKit)
+    /// Selection mode (replace/add/subtract) captured when a polygon draft begins.
+    var polygonSelectionCommitMode: SelectionMode?
+    #endif
     var fillContextMaskScaleUndoBaseline: Double?
 
     /// Target generation budget in megapixels (the *maximum* total pixels). The
@@ -150,12 +154,12 @@ class ImageGenerationViewModel: ObservableObject {
     /// so two generations never touch the pipeline at once. `isGenerating` is
     /// the user-facing flag (cleared the instant Cancel is pressed); this one
     /// trails it until the abandoned compute actually releases.
-    @Published private(set) var isPipelineBusy: Bool = false
+    @Published var isPipelineBusy: Bool = false
     @Published var currentStep: Int = 0
     @Published var totalSteps: Int = 0
     @Published var generatedImage: CGImage?
     /// Formatted input scaled to match the latest output dimensions for aligned A/B preview.
-    @Published private(set) var formattedComparisonImage: CGImage?
+    @Published var formattedComparisonImage: CGImage?
     @Published var previewZoomScale: Double = 1.0
     @Published var previewComparisonSide: PreviewComparisonSide = .processed
     @Published var inputSaveSource: ImageInputSaveSource = .formatted
@@ -165,7 +169,7 @@ class ImageGenerationViewModel: ObservableObject {
 
     /// URL of the most recently saved generated image. Drives "Open Folder" and
     /// names the companion "-input" file. Session-only; never persisted.
-    @Published private(set) var lastSavedImageURL: URL?
+    @Published var lastSavedImageURL: URL?
 
     /// Window title: the loaded project's file name (no extension), or a
     /// placeholder when nothing has been saved/opened yet.
@@ -197,13 +201,13 @@ class ImageGenerationViewModel: ObservableObject {
     @Published var showCheckpoints: Bool = true
 
     // MARK: - Pipeline
-    private var pipeline: Flux2Pipeline?
-    private var generationTask: Task<Void, Never>?
+    var pipeline: Flux2Pipeline?
+    var generationTask: Task<Void, Never>?
     var skipNextModelDefaultApplication = false
 
     // MARK: - Init with defaults
-    private let loadsEnvironmentProject: Bool
-    private let workflow: GenerationWorkflow
+    let loadsEnvironmentProject: Bool
+    let workflow: GenerationWorkflow
 
     init(loadsEnvironmentProject: Bool = false, workflow: GenerationWorkflow = .textToImage) {
         self.loadsEnvironmentProject = loadsEnvironmentProject
@@ -297,25 +301,6 @@ class ImageGenerationViewModel: ObservableObject {
         hasLocalFillSelection
             && enrichInpaintPromptWithVLM
             && abs(fillContextMaskScale) > 0.0001
-    }
-
-    func isMaskToolEnabled(_ tool: InpaintMaskTool) -> Bool {
-        guard canUseCanvasTools else { return false }
-        switch tool {
-        case .liveArea:
-            return barnDoorToolsApply
-        case .rectangle, .polygon, .visionSubject, .cropCanvas:
-            return true
-        case .pointer:
-            return false
-        }
-    }
-
-    var hasActiveSelection: Bool {
-        !inpaintMaskLayers.isEmpty
-            || !draftPolygonPoints.isEmpty
-            || !draftLassoPoints.isEmpty
-            || processArea != nil
     }
 
     /// Mistral text-only upsampling — full-image I2I and outpaint, not masked fill.
@@ -427,72 +412,6 @@ class ImageGenerationViewModel: ObservableObject {
         return "Selection: none"
     }
 
-    var hasFillMask: Bool {
-        !inpaintMaskLayers.isEmpty
-    }
-
-    var canUndoSelection: Bool {
-        workflow == .imageToImage && selectionUndoStore.canUndo
-    }
-
-    var canRedoSelection: Bool {
-        workflow == .imageToImage && selectionUndoStore.canRedo
-    }
-
-    func undoSelection() {
-        guard workflow == .imageToImage else { return }
-        guard let snapshot = selectionUndoStore.popUndo(current: captureSelectionSnapshot()) else { return }
-        restoreSelectionSnapshot(snapshot)
-    }
-
-    func redoSelection() {
-        guard workflow == .imageToImage else { return }
-        guard let snapshot = selectionUndoStore.popRedo(current: captureSelectionSnapshot()) else { return }
-        restoreSelectionSnapshot(snapshot)
-    }
-
-    func clearSelectionUndoHistory() {
-        selectionUndoStore.reset()
-        objectWillChange.send()
-    }
-
-    private func captureSelectionSnapshot() -> SelectionUndoSnapshot {
-        SelectionUndoSnapshot(
-            inpaintMaskLayers: inpaintMaskLayers,
-            processArea: processArea,
-            draftPolygonPoints: draftPolygonPoints,
-            draftLassoPoints: draftLassoPoints,
-            fillContextMaskScale: fillContextMaskScale,
-            inpaintIntent: inpaintIntent,
-            enrichInpaintPromptWithVLM: enrichInpaintPromptWithVLM
-        )
-    }
-
-    func recordSelectionUndoPoint() {
-        guard workflow == .imageToImage, !isApplyingSelectionUndo else { return }
-        selectionUndoStore.pushUndoPoint(captureSelectionSnapshot())
-        objectWillChange.send()
-    }
-
-    private func restoreSelectionSnapshot(_ snapshot: SelectionUndoSnapshot) {
-        isApplyingSelectionUndo = true
-        defer { isApplyingSelectionUndo = false }
-
-        inpaintMaskLayers = snapshot.inpaintMaskLayers
-        processArea = snapshot.processArea
-        draftPolygonPoints = snapshot.draftPolygonPoints
-        draftLassoPoints = snapshot.draftLassoPoints
-        fillContextMaskScale = snapshot.fillContextMaskScale
-        inpaintIntent = snapshot.inpaintIntent
-        enrichInpaintPromptWithVLM = snapshot.enrichInpaintPromptWithVLM
-        visionSubjectMasks.removeAll()
-        visionSubjectStatusMessage = nil
-        isDrawingSelection = false
-        objectWillChange.send()
-
-        Task { await refreshVisionSubjectMaskCache() }
-    }
-
     /// How much the resolution cap shrinks the working image before generative fill.
     var fillDownscaleFactor: Double {
         guard let image = primaryReferenceImage else { return 1 }
@@ -523,12 +442,12 @@ class ImageGenerationViewModel: ObservableObject {
         return "Canvas: \(size.width)×\(size.height) (\(String(format: "%.2f", mp)) MP) — +L\(outpaintPadding.left) +R\(outpaintPadding.right) +T\(outpaintPadding.top) +B\(outpaintPadding.bottom)"
     }
 
-    private static let lastProjectURLKey = "lastGenerationProjectURL"
+    static let lastProjectURLKey = "lastGenerationProjectURL"
     /// Absolute path to a generation project JSON file. When set, opens on launch
     /// instead of restoring `lastGenerationProjectURL` (VM smoke / agent hooks).
-    private static let projectEnvironmentKey = "F2SM_PROJECT"
+    static let projectEnvironmentKey = "F2SM_PROJECT"
     /// When set alongside `F2SM_PROJECT`, writes `ok` or `error` plus detail on load.
-    private static let smokeMarkerEnvironmentKey = "F2SM_SMOKE_MARKER"
+    static let smokeMarkerEnvironmentKey = "F2SM_SMOKE_MARKER"
 
     // MARK: - Image Management
 
@@ -703,264 +622,6 @@ class ImageGenerationViewModel: ObservableObject {
         processArea = Self.clampUnitRect(rect)
     }
 
-    func selectMaskTool(_ tool: InpaintMaskTool) {
-        guard isMaskToolEnabled(tool) else { return }
-        if tool == .cropCanvas {
-            deselectSelections()
-            resetOutpaintCanvas()
-            cancelBarnDoorsIfActive()
-        }
-        inpaintMaskTool = tool
-    }
-
-    /// Drop back to pointer when the active toolbar tool is no longer usable.
-    func clearActiveToolIfDisabled() {
-        guard inpaintMaskTool != .pointer else { return }
-        if !isMaskToolEnabled(inpaintMaskTool) {
-            inpaintMaskTool = .pointer
-        }
-    }
-
-    func deselectSelections(recordUndo: Bool = true) {
-        if recordUndo, hasActiveSelection {
-            recordSelectionUndoPoint()
-        }
-        inpaintMaskLayers.removeAll()
-        draftPolygonPoints.removeAll()
-        draftLassoPoints.removeAll()
-        #if canImport(AppKit)
-        polygonSelectionCommitMode = nil
-        #endif
-        visionSubjectMasks.removeAll()
-        visionSubjectStatusMessage = nil
-        processArea = nil
-        isDrawingSelection = false
-        fillContextMaskScale = 0
-        inpaintMaskTool = .pointer
-    }
-
-    private func applyGenerativeFillDefaultsIfNeeded() {
-        guard hasLocalFillSelection else { return }
-        enrichInpaintPromptWithVLM = true
-        inpaintIntent = .fill
-        fillContextMaskScale = 0
-    }
-
-    #if canImport(AppKit)
-    static func selectionModeFromModifierFlags(_ flags: NSEvent.ModifierFlags) -> SelectionMode {
-        if flags.contains(.option) { return .subtract }
-        if flags.contains(.shift) { return .add }
-        return .replace
-    }
-
-    private static var currentSelectionCommitMode: SelectionMode {
-        selectionModeFromModifierFlags(NSEvent.modifierFlags)
-    }
-    #else
-    private static var currentSelectionCommitMode: SelectionMode { .replace }
-    #endif
-
-    private func prepareSelectionCommit(mode: SelectionMode) {
-        if mode == .replace {
-            inpaintMaskLayers.removeAll()
-            visionSubjectMasks.removeAll()
-            visionSubjectStatusMessage = nil
-        }
-    }
-
-    private func appendCommittedLayer(_ primitive: InpaintMaskPrimitive, mode: SelectionMode) {
-        let hadCommittedSelection = !inpaintMaskLayers.isEmpty
-        recordSelectionUndoPoint()
-        prepareSelectionCommit(mode: mode)
-        inpaintMaskLayers.append(
-            InpaintMaskLayer(
-                combineMode: mode.combineMode,
-                primitive: primitive
-            )
-        )
-        processArea = nil
-        if !hadCommittedSelection {
-            applyGenerativeFillDefaultsIfNeeded()
-        }
-    }
-
-    func commitFillRectangle(_ rect: CGRect, mode: SelectionMode? = nil) {
-        guard inpaintMaskTool != .visionSubject else { return }
-        let resolvedMode = mode ?? Self.currentSelectionCommitMode
-        let clamped = Self.clampUnitRect(rect)
-        appendCommittedLayer(.rectangle(.init(clamped)), mode: resolvedMode)
-    }
-
-    #if canImport(AppKit)
-    private var polygonSelectionCommitMode: SelectionMode?
-    #endif
-
-    func addDraftPolygonPoint(_ point: CGPoint) {
-        #if canImport(AppKit)
-        if draftPolygonPoints.isEmpty {
-            polygonSelectionCommitMode = Self.currentSelectionCommitMode
-        }
-        #endif
-        draftPolygonPoints.append(Self.clampUnitPoint(point))
-    }
-
-    func closeDraftPolygon(mode: SelectionMode? = nil) {
-        guard draftPolygonPoints.count >= 3 else { return }
-        let points = draftPolygonPoints.map { FluxGenerationProject.NormalizedPoint($0) }
-        #if canImport(AppKit)
-        let resolvedMode = mode ?? polygonSelectionCommitMode ?? Self.currentSelectionCommitMode
-        polygonSelectionCommitMode = nil
-        #else
-        let resolvedMode = mode ?? Self.currentSelectionCommitMode
-        #endif
-        if inpaintMaskTool == .visionSubject {
-            commitVisionSubject(selection: .polygon(points), mode: resolvedMode)
-            return
-        }
-        appendCommittedLayer(.polygon(points), mode: resolvedMode)
-        draftPolygonPoints.removeAll()
-    }
-
-    func commitLassoSelection(mode: SelectionMode? = nil) {
-        guard inpaintMaskTool == .visionSubject, draftLassoPoints.count >= 3 else { return }
-        let resolvedMode = mode ?? Self.currentSelectionCommitMode
-        let points = draftLassoPoints.map { FluxGenerationProject.NormalizedPoint($0) }
-        draftLassoPoints.removeAll()
-        isDrawingSelection = false
-        commitVisionSubject(selection: .polygon(points), mode: resolvedMode)
-    }
-
-    func appendLassoPoint(_ point: CGPoint) {
-        draftLassoPoints.append(Self.clampUnitPoint(point))
-    }
-
-    func commitVisionSubject(
-        selection: VisionSubjectSelection,
-        mode: SelectionMode? = nil
-    ) {
-        guard let image = primaryReferenceImage else { return }
-        let mode = mode ?? Self.currentSelectionCommitMode
-        Task {
-            await resolveVisionSubjectLayer(selection: selection, image: image, selectionMode: mode)
-        }
-    }
-
-    @MainActor
-    private func resolveVisionSubjectLayer(
-        selection: VisionSubjectSelection,
-        image: CGImage,
-        selectionMode: SelectionMode
-    ) async {
-        visionSubjectStatusMessage = "Finding subject…"
-        let layerID = UUID()
-        do {
-            let mask = try await Task.detached(priority: .userInitiated) { [inpaintIntent] in
-                try Self.resolveVisionSubjectMask(
-                    selection: selection,
-                    image: image,
-                    inpaintIntent: inpaintIntent
-                )
-            }.value
-            let hadCommittedSelection = !inpaintMaskLayers.isEmpty
-            recordSelectionUndoPoint()
-            if selectionMode == .replace {
-                inpaintMaskLayers.removeAll()
-                visionSubjectMasks.removeAll()
-                visionSubjectStatusMessage = nil
-            }
-            inpaintMaskLayers.append(
-                InpaintMaskLayer(
-                    id: layerID,
-                    combineMode: selectionMode.combineMode,
-                    primitive: .visionSubject(selection)
-                )
-            )
-            visionSubjectMasks[layerID] = mask
-            processArea = nil
-            draftPolygonPoints.removeAll()
-            visionSubjectStatusMessage = nil
-            if !hadCommittedSelection {
-                applyGenerativeFillDefaultsIfNeeded()
-            }
-        } catch {
-            processArea = nil
-            draftPolygonPoints.removeAll()
-            visionSubjectStatusMessage = "No subject in selection"
-        }
-    }
-
-    nonisolated private static func resolveVisionSubjectMask(
-        selection: VisionSubjectSelection,
-        image: CGImage,
-        inpaintIntent: Flux2InpaintIntent
-    ) throws -> CGImage {
-        if #available(macOS 14.0, *) {
-            let region = selection.visionRegion
-            if inpaintIntent == .changeScene {
-                return try Flux2SubjectMask.pickChangeSceneMask(from: image, region: region)
-            }
-            return try Flux2SubjectMask.pickSubjectInpaintMask(from: image, region: region)
-        }
-        throw Flux2Error.invalidConfiguration("Apple Vision subject detection requires macOS 14 or later.")
-    }
-
-    @MainActor
-    func refreshVisionSubjectMaskCache() async {
-        guard let image = primaryReferenceImage else { return }
-        for layer in inpaintMaskLayers {
-            guard case .visionSubject(let selection) = layer.primitive else { continue }
-            if let cached = visionSubjectMasks[layer.id],
-               cached.width == image.width,
-               cached.height == image.height {
-                continue
-            }
-            do {
-                let mask = try await Task.detached(priority: .userInitiated) { [inpaintIntent] in
-                    try Self.resolveVisionSubjectMask(
-                        selection: selection,
-                        image: image,
-                        inpaintIntent: inpaintIntent
-                    )
-                }.value
-                visionSubjectMasks[layer.id] = mask
-            } catch {
-                continue
-            }
-        }
-    }
-
-    func clearFillMask() {
-        deselectSelections()
-    }
-
-    func clearProcessSelection() {
-        deselectSelections()
-    }
-
-    func buildGenerativeFillMask(for image: CGImage) throws -> CGImage {
-        var visionMasks = visionSubjectMasks
-        for layer in inpaintMaskLayers {
-            guard case .visionSubject(let selection) = layer.primitive else { continue }
-            if let cached = visionMasks[layer.id],
-               cached.width == image.width,
-               cached.height == image.height {
-                continue
-            }
-            visionMasks[layer.id] = try Self.resolveVisionSubjectMask(
-                selection: selection,
-                image: image,
-                inpaintIntent: inpaintIntent
-            )
-        }
-
-        return try ImageMaskBuilder.buildInpaintMask(
-            definition: InpaintMaskDefinition(layers: inpaintMaskLayers),
-            image: image,
-            legacyRectangle: processArea,
-            visionMasks: visionMasks
-        )
-    }
-
     static func clampUnitPoint(_ point: CGPoint) -> CGPoint {
         CGPoint(
             x: min(max(point.x, 0), 1),
@@ -999,14 +660,14 @@ class ImageGenerationViewModel: ObservableObject {
         ImagePreparation.conditioningPixelBudget(for: megapixelBudget)
     }
 
-    private func currentPreparationSettings() -> ImagePreparationSettings {
+    func currentPreparationSettings() -> ImagePreparationSettings {
         guard let primary = primaryImageSlot, primary.role == .reference else {
             return ImagePreparationSettings()
         }
         return slotPreparationSettings(for: primary, includeLiveArea: true)
     }
 
-    private func prepareImageToImageInput() throws -> PreparedImageToImageInput {
+    func prepareImageToImageInput() throws -> PreparedImageToImageInput {
         let refs = orderedReferenceSlots
         guard let primary = refs.first, let primaryImage = primary.image else {
             throw Flux2Error.invalidConfiguration("Add a primary reference image before generating")
@@ -1037,7 +698,7 @@ class ImageGenerationViewModel: ObservableObject {
         )
     }
 
-    private func compositeGeneratedImage(_ generatedImage: CGImage, using plan: ImageCompositionPlan) throws -> CGImage {
+    func compositeGeneratedImage(_ generatedImage: CGImage, using plan: ImageCompositionPlan) throws -> CGImage {
         try ImagePreparation.composite(generatedImage, using: plan)
     }
 
@@ -1083,959 +744,6 @@ class ImageGenerationViewModel: ObservableObject {
         return thumbnail
     }
 
-    // MARK: - Projects
-
-    func newProject() {
-        selectedFamily = .flux2
-        selectedModel = .klein4B
-        textQuantization = .mlx8bit
-        transformerQuantization = .qint8
-        prompt = ""
-        upsamplePrompt = false
-        clearPromptAfterGeneration = false
-        upsampledPrompt = nil
-        width = 1024
-        height = 1024
-        seed = ""
-        megapixelBudget = 1.0
-        preparationOverlayOpacity = 0.22
-        processArea = nil
-        contextArea = CGRect(x: 0, y: 0, width: 1, height: 1)
-        clearAllImageSlots()
-        generatedImage = nil
-        checkpointImages.removeAll()
-        errorMessage = nil
-        currentProjectURL = nil
-        lastSavedImageURL = nil
-        clearEditHistory()
-        UserDefaults.standard.removeObject(forKey: Self.lastProjectURLKey)
-        applyRecommendedDefaults(for: selectedModel)
-        ImageSavePreferenceKeys.applyStoredDefaultsToWorking()
-        statusMessage = "New project"
-    }
-
-    func saveProject() {
-        do {
-            let url: URL
-            if let currentProjectURL, !Flux2ProjectDocument.isLegacyJSONProjectURL(currentProjectURL) {
-                url = currentProjectURL
-            } else {
-                let panel = NSSavePanel()
-                panel.allowedContentTypes = Flux2ProjectDocument.saveContentTypes
-                panel.nameFieldStringValue = currentProjectURL.map {
-                    Flux2ProjectDocument.normalizedBundleURL(from: $0).lastPathComponent
-                } ?? Flux2ProjectDocument.defaultSaveName
-                guard panel.runModal() == .OK, let selectedURL = panel.url else {
-                    return
-                }
-                url = Flux2ProjectDocument.normalizedBundleURL(from: selectedURL)
-            }
-
-            try saveProject(to: url)
-            currentProjectURL = url
-            UserDefaults.standard.set(url.path, forKey: Self.lastProjectURLKey)
-            statusMessage = "Saved project to \(url.lastPathComponent)"
-        } catch {
-            errorMessage = "Failed to save project: \(error.localizedDescription)"
-        }
-    }
-
-    func saveProjectAs() {
-        let panel = NSSavePanel()
-        panel.allowedContentTypes = Flux2ProjectDocument.saveContentTypes
-        let suggested = currentProjectURL.map { Flux2ProjectDocument.normalizedBundleURL(from: $0).lastPathComponent }
-            ?? Flux2ProjectDocument.defaultSaveName
-        panel.nameFieldStringValue = suggested
-
-        guard panel.runModal() == .OK, let url = panel.url else {
-            return
-        }
-
-        do {
-            try saveProject(to: Flux2ProjectDocument.normalizedBundleURL(from: url))
-            currentProjectURL = Flux2ProjectDocument.normalizedBundleURL(from: url)
-            UserDefaults.standard.set(currentProjectURL?.path, forKey: Self.lastProjectURLKey)
-            statusMessage = "Saved project to \(currentProjectURL?.lastPathComponent ?? url.lastPathComponent)"
-        } catch {
-            errorMessage = "Failed to save project: \(error.localizedDescription)"
-        }
-    }
-
-    func openProject() {
-        let panel = NSOpenPanel()
-        panel.allowedContentTypes = Flux2ProjectDocument.allowedOpenContentTypes
-        panel.allowsMultipleSelection = false
-        panel.canChooseFiles = true
-        panel.canChooseDirectories = true
-        panel.treatsFilePackagesAsDirectories = true
-
-        guard panel.runModal() == .OK, let url = panel.url else {
-            return
-        }
-
-        do {
-            try loadProject(from: url)
-            currentProjectURL = resolvedProjectURL(from: url)
-            UserDefaults.standard.set(currentProjectURL?.path, forKey: Self.lastProjectURLKey)
-            statusMessage = "Opened project \(currentProjectURL?.lastPathComponent ?? url.lastPathComponent)"
-        } catch {
-            errorMessage = "Failed to open project: \(error.localizedDescription)"
-        }
-    }
-
-    private func loadStartupProjectIfAvailable() {
-        if let envPath = ProcessInfo.processInfo.environment[Self.projectEnvironmentKey],
-           !envPath.isEmpty {
-            loadProjectFromEnvironment(path: envPath)
-            return
-        }
-        loadLastProjectIfAvailable()
-    }
-
-    private func loadProjectFromEnvironment(path: String) {
-        let url = URL(fileURLWithPath: path)
-        guard FileManager.default.fileExists(atPath: url.path) else {
-            let message = "F2SM_PROJECT file not found: \(path)"
-            errorMessage = message
-            writeSmokeMarker(outcome: "error", detail: message)
-            return
-        }
-
-        do {
-            try loadProject(from: url)
-            currentProjectURL = resolvedProjectURL(from: url)
-            UserDefaults.standard.set(currentProjectURL?.path, forKey: Self.lastProjectURLKey)
-            statusMessage = "Opened project \(url.lastPathComponent) (F2SM_PROJECT)"
-            writeSmokeMarker(
-                outcome: "ok",
-                detail: """
-                project=\(url.path)
-                references=\(assignedReferenceCount)
-                prompt=\(prompt)
-                \(editHistorySmokeSummary())
-                """
-            )
-        } catch {
-            let message = "Failed to open F2SM_PROJECT: \(error.localizedDescription)"
-            errorMessage = message
-            currentProjectURL = nil
-            writeSmokeMarker(outcome: "error", detail: message)
-        }
-    }
-
-    private func writeSmokeMarker(outcome: String, detail: String) {
-        guard let markerPath = ProcessInfo.processInfo.environment[Self.smokeMarkerEnvironmentKey],
-              !markerPath.isEmpty else {
-            return
-        }
-
-        let body = "\(outcome)\n\(detail)\n"
-        do {
-            try body.write(toFile: markerPath, atomically: true, encoding: .utf8)
-        } catch {
-            NSLog("F2SM_SMOKE_MARKER write failed: \(error.localizedDescription)")
-        }
-    }
-
-    private func loadLastProjectIfAvailable() {
-        guard let path = UserDefaults.standard.string(forKey: Self.lastProjectURLKey),
-              !path.isEmpty else {
-            return
-        }
-
-        let url = URL(fileURLWithPath: path)
-        guard FileManager.default.fileExists(atPath: url.path) else {
-            return
-        }
-
-        do {
-            try loadProject(from: url)
-            currentProjectURL = resolvedProjectURL(from: url)
-            statusMessage = "Opened last project \(currentProjectURL?.lastPathComponent ?? url.lastPathComponent)"
-        } catch {
-            // Last-project restore should not block a fresh app launch.
-            currentProjectURL = nil
-        }
-    }
-
-    private func saveProject(to url: URL) throws {
-        let bundleURL = Flux2ProjectDocument.normalizedBundleURL(from: url)
-        let project = try makeProject(forBundle: true)
-        let slotImages = imageSlots.compactMap { slot -> FluxGenerationProjectBundle.SlotImage? in
-            guard slot.hasImage, let image = slot.image else { return nil }
-            return FluxGenerationProjectBundle.SlotImage(id: slot.id, image: image)
-        }
-        try FluxGenerationProjectBundle.save(
-            project: project,
-            slotImages: slotImages,
-            previewImage: previewDisplayImage,
-            historyAssets: try historyAssetsForSave(),
-            to: bundleURL
-        )
-    }
-
-    private func resolvedProjectURL(from url: URL) -> URL {
-        if FluxGenerationProjectBundle.isBundleURL(url) {
-            return url.pathExtension == FluxGenerationProjectBundle.packageExtension
-                ? url
-                : url.deletingLastPathComponent()
-        }
-        return url
-    }
-
-    private func loadProject(from url: URL) throws {
-        lastSavedImageURL = nil
-
-        if url.pathExtension == "json" || url.lastPathComponent == FluxGenerationProjectBundle.manifestName {
-            let data = try Data(contentsOf: url.pathExtension == "json" ? url : FluxGenerationProjectBundle.manifestURL(in: url))
-            if let object = try JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let version = object["version"] as? Int,
-               version < FluxGenerationProject.minimumLoadableVersion {
-                let legacy = try JSONDecoder().decode(FluxGenerationProjectV1Legacy.self, from: data)
-                applyProjectV1Shell(legacy)
-                clearAllImageSlots()
-                generatedImage = nil
-                formattedComparisonImage = nil
-                upsampledPrompt = nil
-                checkpointImages.removeAll()
-                errorMessage = nil
-                statusMessage = "This project used the old image format (v1). Images were cleared — add them again in the Images palette."
-                clearSelectionUndoHistory()
-                clearEditHistory()
-                applySizingControls()
-                return
-            }
-        }
-
-        let loaded = try FluxGenerationProject.load(at: url)
-        applyProjectShell(from: loaded.project)
-        try restoreImageSlots(from: loaded.project, bundleRoot: loaded.bundleRoot)
-        loadEditHistory(from: loaded.project, bundleRoot: loaded.bundleRoot)
-        Task { await refreshVisionSubjectMaskCache() }
-        generatedImage = loaded.previewImage
-        formattedComparisonImage = nil
-        upsampledPrompt = nil
-        checkpointImages.removeAll()
-        errorMessage = nil
-        clearSelectionUndoHistory()
-        applySizingControls()
-        applyLoadedHistoryPointer(bundleRoot: loaded.bundleRoot)
-    }
-
-    private func makeProject(forBundle: Bool = false) throws -> FluxGenerationProject {
-        let records: [GenerationImageRecord]
-        if forBundle {
-            records = imageSlots.map { slot in
-                let bundlePath = slot.hasImage
-                    ? FluxGenerationProjectBundle.slotRelativePath(for: slot.id)
-                    : nil
-                var record = slot.toProjectRecord(bundlePath: bundlePath)
-                record.sourcePath = nil
-                record.pngBase64 = nil
-                return record
-            }
-        } else {
-            records = try imageSlots.map { slot in
-                let pngBase64: String?
-                if slot.hasImage, let image = slot.image {
-                    pngBase64 = try pngData(from: image).base64EncodedString()
-                } else {
-                    pngBase64 = nil
-                }
-                return slot.toProjectRecord(pngBase64: pngBase64)
-            }
-        }
-
-        return FluxGenerationProject(
-            version: forBundle ? FluxGenerationProject.bundleVersion : FluxGenerationProject.minimumLoadableVersion,
-            selectedModel: selectedModel.rawValue,
-            textQuantization: textQuantization.rawValue,
-            transformerQuantization: transformerQuantization.rawValue,
-            prompt: prompt,
-            upsamplePrompt: upsamplePrompt,
-            width: width,
-            height: height,
-            steps: steps,
-            guidance: guidance,
-            seed: seed,
-            preparationOverlayOpacity: preparationOverlayOpacity,
-            megapixelBudget: megapixelBudget,
-            clearPromptAfterGeneration: clearPromptAfterGeneration,
-            selectedFamily: selectedFamily?.rawValue,
-            processArea: processArea.map(FluxGenerationProject.NormalizedRect.init),
-            contextArea: FluxGenerationProject.NormalizedRect(contextArea),
-            editMode: nil,
-            inpaintMaskTool: inpaintMaskTool.rawValue,
-            outpaintPadding: outpaintPadding.hasExpansion ? outpaintPadding : nil,
-            inpaintIntent: inpaintIntent.rawValue,
-            enrichInpaintPromptWithVLM: enrichInpaintPromptWithVLM,
-            fillContextMaskScale: hasLocalFillSelection ? fillContextMaskScale : nil,
-            inpaintMaskLayers: inpaintMaskLayers.isEmpty ? nil : inpaintMaskLayers,
-            images: records,
-            selectedImageSlotID: selectedImageSlotID,
-            currentHistoryIndex: editHistoryManifestFields().currentIndex,
-            history: editHistoryManifestFields().entries
-        )
-    }
-
-    func pngData(from image: CGImage) throws -> Data {
-        let data = NSMutableData()
-        guard let destination = CGImageDestinationCreateWithData(data, "public.png" as CFString, 1, nil) else {
-            throw NSError(domain: "FluxProject", code: 2, userInfo: [NSLocalizedDescriptionKey: "Could not encode reference image"])
-        }
-        CGImageDestinationAddImage(destination, image, nil)
-        guard CGImageDestinationFinalize(destination) else {
-            throw NSError(domain: "FluxProject", code: 3, userInfo: [NSLocalizedDescriptionKey: "Could not finalize reference image"])
-        }
-        return data as Data
-    }
-
-    // MARK: - Generation
-
-    /// Generate image (T2I or I2I based on reference images).
-    ///
-    /// Lifecycle (start guard + `isGenerating`) is owned by `startGeneration()`.
-    /// The `defer` always releases the pipeline, even on cancel or error.
-    func generate() async {
-        defer {
-            isGenerating = false
-            isPipelineBusy = false
-        }
-
-        errorMessage = nil
-        generatedImage = nil
-        upsampledPrompt = nil
-        checkpointImages.removeAll()
-        currentStep = 0
-        totalSteps = steps
-        statusMessage = "Initializing pipeline..."
-
-        do {
-            // Create quantization config
-            let quantConfig = Flux2QuantizationConfig(
-                textEncoder: textQuantization,
-                transformer: transformerQuantization
-            )
-
-            // Get HF token
-            let hfToken = ProcessInfo.processInfo.environment["HF_TOKEN"]
-                ?? UserDefaults.standard.string(forKey: "hfToken")
-
-            // Create pipeline
-            statusMessage = "Creating pipeline for \(selectedModel.displayName)..."
-            pipeline = Flux2Pipeline(
-                model: selectedModel,
-                quantization: quantConfig,
-                hfToken: hfToken
-            )
-
-            // Load models
-            statusMessage = "Loading models..."
-            try await pipeline!.loadModels { progress, message in
-                Task { @MainActor in
-                    guard self.isGenerating else { return }
-                    self.statusMessage = message
-                }
-            }
-
-            // Honor a cancel requested during loading before doing expensive work.
-            try Task.checkCancellation()
-
-            // Generate
-            var image: CGImage
-            let interpretPaths = try interpretPathsForGeneration()
-
-            // Checkpoint callback
-            var checkpointCallback: (@Sendable (Int, CGImage) -> Void)? = nil
-            if showCheckpoints {
-                checkpointCallback = { [weak self] step, checkpointImage in
-                    Task { @MainActor in
-                        guard let self, self.isGenerating else { return }
-                        self.addCheckpoint(image: checkpointImage, step: step)
-                    }
-                }
-            }
-
-            // Surface the enhanced prompt the moment the VLM returns it, before
-            // the long denoise, so the UI shows it live during generation.
-            let promptUpsampledCallback: Flux2PromptUpsampleCallback = { [weak self] enhanced in
-                Task { @MainActor in
-                    guard let self, self.isGenerating else { return }
-                    self.upsampledPrompt = enhanced
-                }
-            }
-
-            if workflow == .textToImage {
-                // Text-to-Image
-                statusMessage = "Generating image..."
-                let result = try await pipeline!.generateTextToImageWithResult(
-                    prompt: prompt,
-                    interpretImagePaths: interpretPaths.isEmpty ? nil : interpretPaths,
-                    height: height,
-                    width: width,
-                    steps: steps,
-                    guidance: guidance,
-                    seed: seedValue,
-                    upsamplePrompt: upsamplePrompt,
-                    checkpointInterval: showCheckpoints ? checkpointInterval : nil,
-                    onProgress: { current, total in
-                        Task { @MainActor in
-                            guard self.isGenerating else { return }
-                            self.currentStep = current
-                            self.totalSteps = total
-                            self.statusMessage = "Step \(current)/\(total)"
-                        }
-                    },
-                    onCheckpoint: checkpointCallback,
-                    onPromptUpsampled: promptUpsampledCallback
-                )
-                image = result.image
-                if result.wasUpsampled { upsampledPrompt = result.usedPrompt }
-            } else if generateRoute == .localFill {
-                guard hasFillMask else {
-                    throw Flux2Error.invalidConfiguration("Draw a selection on the reference image before generating.")
-                }
-
-                if enrichInpaintPromptWithVLM {
-                    statusMessage = "Loading Qwen3.5 VLM..."
-                    try await ensureQwen35VLMLoaded()
-                }
-
-                guard let sourceImage = primaryReferenceImage else {
-                    throw Flux2Error.invalidConfiguration("Add a primary reference image before generating.")
-                }
-                statusMessage = "Preparing selection..."
-                let mask = try buildGenerativeFillMask(for: sourceImage)
-
-                statusMessage = "Running generative fill..."
-                let maxPixels = Int(megapixelBudget * 1_000_000)
-                let chain = Flux2MaskedInpaintingChain(
-                    pipeline: pipeline!,
-                    prompt: prompt.isEmpty ? "fill in this region" : prompt,
-                    image: sourceImage,
-                    mask: mask,
-                    steps: steps,
-                    guidance: guidance,
-                    seed: seedValue,
-                    upsamplePrompt: upsamplePrompt,
-                    enrichPromptWithVLM: enrichInpaintPromptWithVLM,
-                    intent: inpaintIntent,
-                    vlmContextArea: enrichInpaintPromptWithVLM ? fillVLMContextArea : nil,
-                    maxPixels: maxPixels,
-                    checkpointInterval: showCheckpoints ? checkpointInterval : nil,
-                    onProgress: { current, total in
-                        Task { @MainActor in
-                            guard self.isGenerating else { return }
-                            self.currentStep = current
-                            self.totalSteps = total
-                            self.statusMessage = "Step \(current)/\(total)"
-                        }
-                    },
-                    onCheckpoint: checkpointCallback
-                )
-                let result = try await chain.run()
-                image = result.image
-                if result.wasUpsampled {
-                    upsampledPrompt = result.usedPrompt
-                } else if enrichInpaintPromptWithVLM,
-                          result.usedPrompt != prompt,
-                          !result.usedPrompt.isEmpty {
-                    upsampledPrompt = result.usedPrompt
-                }
-                if let notice = result.notice {
-                    statusMessage = "Generation complete — \(notice)"
-                }
-            } else if generateRoute == .outpaint {
-                guard outpaintCanvasIsDefined else {
-                    throw Flux2Error.invalidConfiguration("Expand the canvas on the preview before generating.")
-                }
-
-                guard let sourceImage = primaryReferenceImage else {
-                    throw Flux2Error.invalidConfiguration("Add a primary reference image before generating.")
-                }
-                statusMessage = "Running outpaint..."
-                let maxPixels = Int(megapixelBudget * 1_000_000)
-                let chain = Flux2OutpaintingChain(
-                    pipeline: pipeline!,
-                    image: sourceImage,
-                    top: outpaintPadding.top,
-                    bottom: outpaintPadding.bottom,
-                    left: outpaintPadding.left,
-                    right: outpaintPadding.right,
-                    prompt: prompt,
-                    steps: steps,
-                    guidance: guidance,
-                    seed: seedValue,
-                    upsamplePrompt: upsamplePrompt,
-                    maxPixels: maxPixels,
-                    onProgress: { current, total in
-                        Task { @MainActor in
-                            guard self.isGenerating else { return }
-                            self.currentStep = current
-                            self.totalSteps = total
-                            self.statusMessage = "Step \(current)/\(total)"
-                        }
-                    }
-                )
-                let result = try await chain.run()
-                image = result.image
-                if result.wasUpsampled { upsampledPrompt = result.usedPrompt }
-                if let notice = result.notice {
-                    statusMessage = "Generation complete — \(notice)"
-                }
-            } else {
-                // Image-to-Image (full-frame + Image Preparation)
-                statusMessage = "Preparing image-to-image input..."
-                let preparedInput = try prepareImageToImageInput()
-
-                if upsamplePrompt {
-                    statusMessage = "Upsampling prompt (Mistral VLM)..."
-                } else {
-                    statusMessage = "Generating with \(preparedInput.images.count) reference image(s)..."
-                }
-                let result = try await pipeline!.generateImageToImageWithResult(
-                    prompt: prompt,
-                    images: preparedInput.images,
-                    interpretImagePaths: interpretPaths.isEmpty ? nil : interpretPaths,
-                    height: preparedInput.height,
-                    width: preparedInput.width,
-                    steps: steps,
-                    guidance: guidance,
-                    seed: seedValue,
-                    upsamplePrompt: upsamplePrompt,
-                    checkpointInterval: showCheckpoints ? checkpointInterval : nil,
-                    onProgress: { current, total in
-                        Task { @MainActor in
-                            guard self.isGenerating else { return }
-                            self.currentStep = current
-                            self.totalSteps = total
-                            self.statusMessage = "Step \(current)/\(total)"
-                        }
-                    },
-                    onCheckpoint: checkpointCallback,
-                    onPromptUpsampled: promptUpsampledCallback
-                )
-                image = result.image
-                if result.wasUpsampled { upsampledPrompt = result.usedPrompt }
-
-                if let compositionPlan = preparedInput.compositionPlan {
-                    statusMessage = "Compositing context area..."
-                    image = try compositeGeneratedImage(image, using: compositionPlan)
-                }
-            }
-
-            // Only publish if this run wasn't cancelled mid-flight. If Cancel was
-            // pressed during the final step, `isGenerating` is already false and
-            // we discard the result rather than flashing it over "Cancelled".
-            if isGenerating {
-                cacheFormattedComparisonImage(for: image)
-                generatedImage = image
-                previewComparisonSide = .processed
-                if generateRoute == .localFill {
-                    fillContextMaskScale = 0
-                }
-                if statusMessage.isEmpty || statusMessage.hasPrefix("Step ") {
-                    statusMessage = "Generation complete!"
-                }
-                if clearPromptAfterGeneration {
-                    prompt = ""
-                }
-                appendEditHistoryAfterGenerate(image: image)
-            }
-
-        } catch is CancellationError {
-            if isGenerating { statusMessage = "Cancelled" }
-        } catch Flux2Error.generationCancelled {
-            if isGenerating { statusMessage = "Cancelled" }
-        } catch {
-            if isGenerating {
-                errorMessage = error.localizedDescription
-                statusMessage = "Generation failed"
-            }
-        }
-    }
-
-    /// Start a cancellable generation run. The Generate button calls this; the
-    /// in-flight Task is owned here so `cancel()` can stop it at any time.
-    func startGeneration() {
-        guard canGenerate else { return }
-        isGenerating = true
-        isPipelineBusy = true
-        generationTask = Task { [weak self] in
-            await self?.generate()
-        }
-    }
-
-    /// Cancel an in-flight generation immediately. The UI is freed right away and
-    /// the in-flight result is discarded; the background diffusion step unwinds on
-    /// its own (the denoising loop checks `Task.checkCancellation()` between steps)
-    /// and releases the pipeline when it does.
-    func cancel() {
-        guard isGenerating else { return }
-        generationTask?.cancel()
-        isGenerating = false
-        currentStep = 0
-        totalSteps = 0
-        statusMessage = "Cancelled"
-    }
-
-    /// Save generated image to file
-    func saveImage() {
-        guard let image = generatedImage else { return }
-
-        do {
-            let url = try ImageSaveService.save(
-                image,
-                metadata: ImageSaveMetadata(prompt: prompt)
-            )
-            lastSavedImageURL = url
-            statusMessage = "Saved to \(url.path)"
-        } catch {
-            errorMessage = "Failed to save: \(error.localizedDescription)"
-        }
-    }
-
-    /// Save the input variant chosen in the Save Input picker (raw / formatted / prepared).
-    func saveInputImage() {
-        do {
-            let (image, suffix) = try resolveInputSaveImage(for: inputSaveSource)
-            let url = try ImageSaveService.save(
-                image,
-                metadata: ImageSaveMetadata(prompt: prompt),
-                stemSuffix: suffix
-            )
-            statusMessage = "Saved \(inputSaveSource.rawValue.lowercased()) input to \(url.lastPathComponent)"
-        } catch {
-            errorMessage = "Failed to save input: \(error.localizedDescription)"
-        }
-    }
-
-    private func resolveInputSaveImage(for source: ImageInputSaveSource) throws -> (CGImage, String?) {
-        switch source {
-        case .raw:
-            guard let image = primaryReferenceImage else {
-                throw Flux2Error.invalidConfiguration("Add a reference image before saving the raw input.")
-            }
-            return (image, "-raw")
-        case .formatted:
-            return (try formattedFullInputImage(), "-formatted")
-        case .prepared:
-            return (try preparedInputImage(), "-prepared")
-        }
-    }
-
-    /// Open the configured output folder in Finder, selecting the last save when known.
-    func openOutputFolder() {
-        #if canImport(AppKit)
-        do {
-            if let url = lastSavedImageURL {
-                try ImageSaveService.revealOutputDirectoryInFinder(selecting: url)
-            } else {
-                try ImageSaveService.revealOutputDirectoryInFinder()
-            }
-        } catch {
-            errorMessage = "Failed to open output folder: \(error.localizedDescription)"
-        }
-        #endif
-    }
-
-    /// Clear the generated image and checkpoints from the preview pane.
-    func clearPreview() {
-        generatedImage = nil
-        formattedComparisonImage = nil
-        previewComparisonSide = .processed
-        checkpointImages.removeAll()
-        statusMessage = "Preview cleared"
-    }
-
-    /// The reference image formatted per Image Formatting (Favour + crop/pad)
-    /// across the *whole* frame — i.e. before the barn doors narrow it. Mirrors
-    /// the output's budget-filled, alignment-floored sizing so the saved input
-    /// shares the model's grid.
-    func formattedFullInputImage() throws -> CGImage {
-        guard let original = primaryReferenceImage else {
-            throw Flux2Error.invalidConfiguration("Add a reference image before saving the input")
-        }
-
-        var settings = currentPreparationSettings()
-        settings.contextArea = CGRect(x: 0, y: 0, width: 1, height: 1)
-        return try ImagePreparation.prepare(referenceImages: [original], settings: settings).images[0]
-    }
-
-    /// The barn-door crop plus formatting — the first reference image sent to the model.
-    func preparedInputImage() throws -> CGImage {
-        guard hasPrimaryReference else {
-            throw Flux2Error.invalidConfiguration("Add a reference image before saving the prepared input")
-        }
-        return try prepareImageToImageInput().images[0]
-    }
-
-    func cacheFormattedComparisonImage(for output: CGImage) {
-        guard let original = primaryReferenceImage else {
-            formattedComparisonImage = nil
-            return
-        }
-
-        formattedComparisonImage = try? ImagePreparation.formatToCanvas(
-            referenceImage: original,
-            settings: currentPreparationSettings(),
-            targetWidth: output.width,
-            targetHeight: output.height
-        )
-    }
-
-    func persistSessionState() {
-        let state = captureGUIState()
-        switch workflow {
-        case .imageToImage:
-            Flux2AppSessionStore.saveImageToImage(state)
-        case .textToImage:
-            Flux2AppSessionStore.saveTextToImage(state)
-        }
-    }
-
-    private func restoreSessionStateIfNeeded() {
-        if loadsEnvironmentProject,
-           let projectPath = ProcessInfo.processInfo.environment["F2SM_PROJECT"],
-           !projectPath.isEmpty {
-            return
-        }
-
-        let state: Flux2GenerationGUIState?
-        switch workflow {
-        case .imageToImage:
-            state = Flux2AppSessionStore.loadImageToImage()
-        case .textToImage:
-            state = Flux2AppSessionStore.loadTextToImage()
-        }
-
-        guard let state else { return }
-        applyGUIState(state, projectLoaded: currentProjectURL != nil)
-    }
-
-    private func captureGUIState() -> Flux2GenerationGUIState {
-        Flux2GenerationGUIState(
-            selectedFamily: selectedFamily?.rawValue,
-            selectedModel: selectedModel.rawValue,
-            textQuantization: textQuantization.rawValue,
-            transformerQuantization: transformerQuantization.rawValue,
-            prompt: prompt,
-            upsamplePrompt: upsamplePrompt,
-            clearPromptAfterGeneration: clearPromptAfterGeneration,
-            width: width,
-            height: height,
-            steps: steps,
-            guidance: guidance,
-            seed: seed,
-            sizingFavor: sizingFavor.rawValue,
-            sizingMethod: sizingMethod.rawValue,
-            preparationScale: preparationScale,
-            preparationOverlayOpacity: preparationOverlayOpacity,
-            megapixelBudget: megapixelBudget,
-            contextAreaX: Double(contextArea.minX),
-            contextAreaY: Double(contextArea.minY),
-            contextAreaWidth: Double(contextArea.width),
-            contextAreaHeight: Double(contextArea.height),
-            processAreaX: processArea.map { Double($0.minX) },
-            processAreaY: processArea.map { Double($0.minY) },
-            processAreaWidth: processArea.map { Double($0.width) },
-            processAreaHeight: processArea.map { Double($0.height) },
-            hasProcessArea: processArea != nil,
-            editMode: nil,
-            inpaintMaskTool: inpaintMaskTool.rawValue,
-            outpaintPadding: outpaintPadding.hasExpansion ? outpaintPadding : nil,
-            inpaintIntent: inpaintIntent.rawValue,
-            enrichInpaintPromptWithVLM: enrichInpaintPromptWithVLM,
-            fillContextMaskScale: hasLocalFillSelection ? fillContextMaskScale : 0,
-            interpretImagePaths: (try? interpretPathsForGeneration()) ?? [],
-            showCheckpoints: showCheckpoints,
-            checkpointInterval: checkpointInterval,
-            previewZoomScale: previewZoomScale,
-            previewComparisonSide: previewComparisonSide.rawValue,
-            inputSaveSource: inputSaveSource.rawValue
-        )
-    }
-
-    private func applyGUIState(_ state: Flux2GenerationGUIState, projectLoaded: Bool) {
-        if !projectLoaded {
-            skipNextModelDefaultApplication = true
-            if let family = state.selectedFamily.flatMap(ModelFamily.init(rawValue:)) {
-                selectedFamily = family
-            }
-            selectedModel = Flux2Model(rawValue: state.selectedModel) ?? selectedModel
-            textQuantization = MistralQuantization(rawValue: state.textQuantization) ?? textQuantization
-            transformerQuantization = TransformerQuantization(rawValue: state.transformerQuantization) ?? transformerQuantization
-            prompt = state.prompt
-            upsamplePrompt = state.upsamplePrompt
-            clearPromptAfterGeneration = state.clearPromptAfterGeneration
-            width = state.width
-            height = state.height
-            steps = state.steps
-            guidance = state.guidance
-            seed = state.seed
-
-            if workflow == .imageToImage {
-                bootstrapImageSlotsIfNeeded()
-                if let index = imageSlots.indices.first {
-                    if let favor = state.sizingFavor.flatMap(ImageSizingFavor.init(rawValue:)) {
-                        imageSlots[index].sizingFavor = favor
-                    }
-                    if let method = state.sizingMethod.flatMap(ImageSizingMethod.init(rawValue:)) {
-                        imageSlots[index].sizingMethod = method
-                    }
-                    if let scale = state.preparationScale {
-                        imageSlots[index].preparationScale = max(0.1, min(1.0, scale))
-                    }
-                }
-                preparationOverlayOpacity = state.preparationOverlayOpacity ?? preparationOverlayOpacity
-                megapixelBudget = state.megapixelBudget ?? megapixelBudget
-                contextArea = Self.clampUnitRect(CGRect(
-                    x: state.contextAreaX ?? contextArea.minX,
-                    y: state.contextAreaY ?? contextArea.minY,
-                    width: state.contextAreaWidth ?? contextArea.width,
-                    height: state.contextAreaHeight ?? contextArea.height
-                ))
-                if state.hasProcessArea,
-                   let x = state.processAreaX,
-                   let y = state.processAreaY,
-                   let width = state.processAreaWidth,
-                   let height = state.processAreaHeight {
-                    processArea = Self.clampUnitRect(CGRect(x: x, y: y, width: width, height: height))
-                } else {
-                    processArea = nil
-                }
-                let legacyRoute = I2IGenerateRoute.fromLegacyProjectValue(state.editMode)
-                if let toolRaw = state.inpaintMaskTool,
-                   let tool = InpaintMaskTool(rawValue: toolRaw) {
-                    inpaintMaskTool = tool
-                } else if legacyRoute == .outpaint {
-                    inpaintMaskTool = .cropCanvas
-                }
-                outpaintPadding = state.outpaintPadding ?? .zero
-                outpaintCanvasIsDefined = outpaintPadding.hasExpansion
-                if outpaintPadding.hasExpansion {
-                    updateOutpaintPadding(outpaintPadding)
-                }
-                inpaintIntent = state.inpaintIntent.flatMap(Flux2InpaintIntent.init(rawValue:)) ?? inpaintIntent
-                enrichInpaintPromptWithVLM = state.enrichInpaintPromptWithVLM ?? enrichInpaintPromptWithVLM
-                fillContextMaskScale = state.fillContextMaskScale
-                applySizingControls()
-            }
-        } else if workflow == .imageToImage {
-            preparationOverlayOpacity = state.preparationOverlayOpacity ?? preparationOverlayOpacity
-        }
-
-        showCheckpoints = state.showCheckpoints
-        checkpointInterval = max(1, state.checkpointInterval)
-        previewZoomScale = min(max(state.previewZoomScale, 0.25), 8.0)
-        previewComparisonSide = PreviewComparisonSide(rawValue: state.previewComparisonSide) ?? .processed
-        if let input = ImageInputSaveSource(rawValue: state.inputSaveSource) {
-            inputSaveSource = input
-        }
-    }
-
-    /// Clear pipeline to free memory
-    func clearPipeline() async {
-        await pipeline?.clearAll()
-        pipeline = nil
-        Memory.clearCache()
-        statusMessage = "Models unloaded"
-    }
-
-    /// Load Qwen3.5 VLM (4-bit) when generative fill requests VLM prompt enrichment.
-    private func ensureQwen35VLMLoaded() async throws {
-        if FluxTextEncoders.shared.isQwen35VLMLoaded { return }
-
-        let hfToken = ProcessInfo.processInfo.environment["HF_TOKEN"]
-            ?? UserDefaults.standard.string(forKey: "hfToken")
-
-        let downloader = TextEncoderModelDownloader(hfToken: hfToken)
-        let path = try await downloader.downloadQwen35(variant: .qwen35_4B_4bit) { progress, message in
-            Task { @MainActor in
-                guard self.isGenerating else { return }
-                self.statusMessage = "Qwen3.5: \(message) (\(Int(progress * 100))%)"
-            }
-        }
-        try await FluxTextEncoders.shared.loadQwen35VLM(from: path.path)
-    }
-
-    // MARK: - Recommended Defaults (Black Forest Labs)
-
-    /// Apply recommended defaults for a model (from official HuggingFace pages)
-    func applyRecommendedDefaults(for model: Flux2Model) {
-        switch model {
-        case .dev:
-            // Flux.2 Dev - https://huggingface.co/black-forest-labs/FLUX.2-dev
-            // 28 steps is "a good trade-off", guidance 4.0
-            textQuantization = .mlx8bit
-            transformerQuantization = .qint8
-            width = 1024
-            height = 1024
-            steps = 28
-            guidance = 4.0
-            checkpointInterval = 7
-
-        case .klein4B, .klein4BBase:
-            // Flux.2 Klein 4B - https://huggingface.co/black-forest-labs/FLUX.2-klein-4B
-            // 4 steps, guidance 1.0, optimized for sub-second generation
-            transformerQuantization = .qint8
-            width = 1024
-            height = 1024
-            steps = 4
-            guidance = 1.0
-            checkpointInterval = 1
-
-        case .klein9B, .klein9BBase, .klein9BKV:
-            // Flux.2 Klein 9B - https://huggingface.co/black-forest-labs/FLUX.2-klein-9B
-            // 4 steps, guidance 1.0, sub-second generation
-            transformerQuantization = .bf16  // Only bf16 available
-            width = 1024
-            height = 1024
-            steps = 4
-            guidance = 1.0
-            checkpointInterval = 1
-        }
-    }
-
-    // MARK: - Presets
-
-    /// Apply a memory-efficient preset (Klein 4B at 512x512)
-    func applyLightweightPreset() {
-        selectedModel = .klein4B
-        applyRecommendedDefaults(for: .klein4B)
-        width = 512
-        height = 512
-    }
-
-    /// Apply a balanced preset (Klein 4B at 1024x1024)
-    func applyBalancedPreset() {
-        selectedModel = .klein4B
-        applyRecommendedDefaults(for: .klein4B)
-    }
-
-    /// Apply a high quality preset (Dev at 1024x1024)
-    func applyHighQualityPreset() {
-        selectedModel = .dev
-        applyRecommendedDefaults(for: .dev)
-    }
-
-    // MARK: - Checkpoints
-
-    /// Clear checkpoint images
-    func clearCheckpoints() {
-        checkpointImages.removeAll()
-    }
-
-    /// Add a checkpoint image
-    func addCheckpoint(image: CGImage, step: Int) {
-        let checkpoint = CheckpointImage(
-            id: UUID(),
-            image: image,
-            step: step,
-            timestamp: Date()
-        )
-        checkpointImages.append(checkpoint)
-    }
 }
 
 // MARK: - Checkpoint Image Model
